Online-ChatMultiPDF/requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-core
huggingface-hub
google-generativeai
groq
python-dotenv
faiss-cpu
requests